MainActivity

+++ app/src/main/java/br/edu/ifsp/ea_2/ui/listeners/DiaryEntryItemClickListener.kt
@@ -2,7 +2,7 @@ package br.edu.ifsp.ea_2.ui.listeners
 
 interface DiaryEntryItemClickListener {
 
-    fun clickOpen(position: Int)
+    fun clickOpen(position: Long)
 
     fun clickDelete(position: Long)
 }
